Start and finish the project
Add the main files and directories, add an image for the
project, add all the HTML code and add all the CSS code.

The image is added as a requirement for the project,
the HTML code fulfills all project requirements, the
CSS code also fulfills all project requirements and
is web responsive.